+++ src/dependency_injector/errors.pyx
@@ -1,7 +1,10 @@
-"""Dependency injector errors module."""
+"""Dependency injector errors.
+
+Powered by Cython.
+"""
 
 
-class Error(Exception):
+cdef class Error(Exception):
     """Base error.
 
     All dependency injector errors extend this error class.